Move float close to view
