+++ src/templates/floating-close-button.js
